Prueba-Tecnica-Ingeniero-Fullstack/fastapi-project/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asarPy==1.0.1
asgiref==3.8.1
bcrypt==4.0.1
black==23.3.0
certifi==2024.12.14
cffi==1.17.1
click==8.1.8
cryptography==44.0.0
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.6
flake8==6.0.0
greenlet==3.1.1
h11==0.14.0
httpcore==0.17.3
httpx==0.24.0
idna==3.10
iniconfig==2.0.0
Mako==1.3.8
MarkupSafe==3.0.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
passlib==1.7.4
pathspec==0.12.1
pipdeptree==2.24.0
platformdirs==4.3.6
pluggy==1.5.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycodestyle==2.10.0
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
pyflakes==3.0.1
pytest==7.3.1
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==1.4.23
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.15.0